--- src/Pages/Services.jsx
@@ -1,18 +1,0 @@
-import React from "react";
-import Navigation from "../components/Sections/Navigation/Navigation";
-import Hero from "../components/Sections/Hero/Hero";
-import FeaturesSection from "../components/Sections/FeaturesSection/FeaturesSection";
-
-
-function Services() {
-	return (
-		<>
-			<Navigation />
-			<Hero />
-			<FeaturesSection />
-			
-		</>
-	);
-}
-
-export default Services;
