--- src/Styles/mixins.scss
@@ -1,3 +1,0 @@
-@mixin centered-text {
-    text-align: center;
-}
