+++ src/components/Navigation/navigation.scss
@@ -1,0 +1,5 @@
+@media screen and (max-width: 768px) {
+    #menu-items {
+        display: none;
+    }
+}
