--- src/components/Sections/Footer/footer.jsx
@@ -1,180 +1,0 @@
-import React from "react";
-import { FaFacebook, FaTwitter, FaInstagram, FaLinkedin } from "react-icons/fa";
-
-function Footer() {
-  return (
-    <section>
-      <div className="mt-8 bg-[#85C1E9] p-6">
-        <div className="w-full sm:flex sm:items-center sm:justify-between">
-          <h2 className="text-[#0C335E] text-xl font-semibold ml-8">
-            Get connected with us on social media
-          </h2>
-          <div className="flex gap-6 sm:mt-0 mt-4 sm:justify-center mr-8">
-            <span style={{ color: "#0C335E" }}>
-              <FaFacebook size={30} />
-            </span>
-            <span style={{ color: "#0C335E" }}>
-              <FaTwitter size={30} />
-            </span>
-            <span style={{ color: "#0C335E" }}>
-              <FaLinkedin size={30} />
-            </span>
-            <span style={{ color: "red" }}>
-              <FaInstagram size={30} />
-            </span>
-          </div>
-        </div>
-      </div>
-      <footer class="bg-[#007BFF] text-white">
-        <div class="container mx-auto px-6 md:flex md:justify-between">
-          {/*  Left Column  */}
-          <div class="mb-8 md:mb-0">
-            <h2 class="text-xl font-bold mb-4">Applora</h2>
-            {/* <p className="line-clamp-*">
-              Applora is designed to provide a comprehensive solution for
-              designers, teams, and businesses to deliver exceptional products.
-            </p> */}
-          </div>
-
-          {/* Services Column  */}
-          <div class="mb-8 md:mb-0">
-            <h3 class="text-xl font-semibold mb-4">SERVICES</h3>
-            <ul>
-              <li>
-                <a href="#" class="hover:underline">
-                  Task Execution
-                </a>
-              </li>
-              <li>
-                <a href="#" class="hover:underline">
-                  Project Execution
-                </a>
-              </li>
-              <li>
-                <a href="#" class="hover:underline">
-                  Invoicing
-                </a>
-              </li>
-              <li>
-                <a href="#" class="hover:underline">
-                  Chat Support
-                </a>
-              </li>
-              <li>
-                <a href="#" class="hover:underline">
-                  Third-Party Tools
-                </a>
-              </li>
-            </ul>
-          </div>
-
-          {/* <!-- Legal Column --> */}
-          <div class="mb-8 md:mb-0">
-            <h3 class="text-xl font-semibold mb-4">LEGAL</h3>
-            <ul>
-              <li>
-                <a href="#" class="hover:underline">
-                  Terms of Service
-                </a>
-              </li>
-              <li>
-                <a href="#" class="hover:underline">
-                  Privacy Policy
-                </a>
-              </li>
-              <li>
-                <a href="#" class="hover:underline">
-                  Cookie Policy
-                </a>
-              </li>
-              <li>
-                <a href="#" class="hover:underline">
-                  Disclaimer
-                </a>
-              </li>
-            </ul>
-          </div>
-
-          {/* Contact Column  */}
-          <div>
-            <h3 class="text-xl font-semibold mb-4">CONTACTS</h3>
-            <ul>
-              <span className="flex">
-                <svg
-                  width="31"
-                  height="31"
-                  viewBox="0 0 31 31"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M17.5697 12.4223V23.2614C17.5697 23.9368 16.6371 24.4904 15.4993 24.4904C14.3615 24.4904 13.4288 23.9478 13.4288 23.2614V12.4223C8.57918 11.8466 4.92329 9.31121 4.92329 6.26652C4.92329 2.81218 9.66103 0 15.4993 0C21.3375 0 26.0566 2.81218 26.0566 6.26652C26.0752 9.31121 22.4007 11.8466 17.5697 12.4223Z"
-                    fill="#0C335E"
-                  />
-                  <path
-                    d="M7.14346 19.7188C8.188 19.4642 9.40041 19.752 9.84807 20.372C10.2957 20.992 9.79211 21.7117 8.74757 21.9774C5.85643 22.6971 4.12175 23.7821 4.12175 24.8892C4.12175 26.6164 8.80353 28.5539 15.4998 28.5539C22.2147 28.5539 26.8778 26.6275 26.8778 24.8892C26.8778 23.7932 25.1058 22.6749 22.2706 21.9774C21.2261 21.7117 20.7225 21.0031 21.1701 20.372C21.6178 19.752 22.8116 19.4531 23.8747 19.7188C28.3513 20.826 31 22.7635 31 24.8892C31 28.3214 24.1918 31.0008 15.4998 31.0008C6.80771 31.0008 -0.00044632 28.3214 -0.00044632 24.8892C-0.00044632 22.7635 2.66686 20.837 7.14346 19.7188Z"
-                    fill="#0C335E"
-                  />
-                </svg>
-                <li class="flex items-center">Lekki Lagos, Nigeria.</li>
-              </span>
-              <span className="flex">
-                <svg
-                  width="25"
-                  height="24"
-                  viewBox="0 0 25 24"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M0.811058 23.5544C0.797972 23.4924 0.789062 23.427 0.789062 23.3582V0.640975C0.789062 0.571704 0.797972 0.505853 0.811337 0.443423C0.813284 0.434871 0.814398 0.426319 0.816626 0.417767C0.834442 0.344647 0.860613 0.278796 0.893465 0.22107C0.895693 0.217222 0.895971 0.212518 0.898197 0.20867C0.89959 0.206532 0.90126 0.205676 0.902653 0.203538C0.937731 0.14624 0.979214 0.100059 1.02571 0.065423C1.02933 0.0628574 1.03211 0.0594364 1.03573 0.0568708C1.07972 0.0265111 1.1276 0.00940733 1.178 0.00427611C1.18301 0.00384851 1.18774 0.00213765 1.19275 0.00213765C1.19748 0.00171005 1.20166 0 1.20639 0H24.3712C24.3759 0 24.3804 0.00171005 24.3851 0.00213765C24.3895 0.00256525 24.394 0.00384851 24.3985 0.00427611C24.4494 0.00940733 24.4981 0.026939 24.5427 0.0577263C24.5455 0.0594367 24.5477 0.0624297 24.5505 0.0641401C24.5978 0.0987758 24.6398 0.145812 24.6752 0.203966C24.6766 0.206104 24.6783 0.206532 24.6796 0.20867C24.6819 0.212091 24.6822 0.216794 24.6844 0.220643C24.7172 0.278369 24.7437 0.344647 24.7615 0.417767C24.7637 0.426319 24.7648 0.434871 24.7668 0.443851C24.7802 0.50628 24.7891 0.572131 24.7891 0.641403V23.359C24.7891 23.4292 24.7799 23.4959 24.7662 23.5591C24.7643 23.5677 24.7632 23.5758 24.7612 23.5844C24.7431 23.6579 24.7164 23.7242 24.6833 23.7819C24.6813 23.7853 24.6808 23.7896 24.6788 23.793C24.6771 23.796 24.6749 23.7977 24.673 23.8007C24.6593 23.8225 24.6443 23.8418 24.6287 23.8606C24.6217 23.8687 24.6156 23.8786 24.6084 23.8863C24.59 23.9059 24.5702 23.9217 24.5499 23.9363C24.5449 23.9397 24.5404 23.9453 24.5354 23.9483C24.485 23.9812 24.4296 24 24.3715 24H1.20695C1.14876 24 1.09336 23.9816 1.04297 23.9483C1.03517 23.9431 1.02794 23.9354 1.02042 23.9294C1.00316 23.9162 0.985897 23.9025 0.970028 23.8858C0.961676 23.8769 0.954157 23.8653 0.946083 23.8555C0.931885 23.8379 0.917965 23.8204 0.905714 23.8003C0.903488 23.7969 0.900982 23.7943 0.898754 23.7909C0.896526 23.7871 0.895971 23.7819 0.893744 23.7781C0.86117 23.7208 0.834999 23.6549 0.817181 23.5822C0.814398 23.5728 0.813007 23.5638 0.811058 23.5544ZM23.9536 2.09568L16.9246 11.9566L23.9536 21.8962V2.09568ZM9.54716 11.4734C9.57612 11.5025 9.60674 11.5269 9.63096 11.5679C9.64015 11.5833 9.64321 11.6034 9.65129 11.6196L12.7889 16.0214L23.2949 1.28238H2.28299L9.54716 11.4734ZM16.3055 12.8251L13.0707 17.3632C12.9908 17.4752 12.89 17.5312 12.7889 17.5312C12.6879 17.5312 12.5871 17.4752 12.5075 17.3632L9.31275 12.8815L2.28494 22.7172H23.301L16.3055 12.8251ZM8.69329 12.0126L1.62428 2.09568V21.9065L8.69329 12.0126Z"
-                    fill="#0C335E"
-                  />
-                </svg>
-
-                <li class="flex items-center">info@applora.com</li>
-              </span>
-              <span className="flex">
-                <svg
-                  width="30"
-                  height="36"
-                  viewBox="0 0 30 36"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M19.6454 10.8507C19.3255 10.5309 19.2445 9.97215 19.6427 9.57403C20.0408 9.17591 20.5195 9.17692 20.9193 9.57672C25.3172 13.9745 25.3323 21.1558 20.953 25.5352C20.6345 25.8537 20.0761 25.9323 19.6763 25.5325C19.2765 25.1327 19.2755 24.6539 19.6736 24.2558C23.4163 20.673 23.4035 14.6088 19.6454 10.8507ZM16.4779 22.3341C16.158 22.0142 16.0771 21.4555 16.4752 21.0574C18.3862 19.1464 18.3798 16.1143 16.4607 14.1953C16.1409 13.8754 16.0599 13.3167 16.458 12.9186C16.8562 12.5205 17.3349 12.5215 17.7347 12.9213C20.2934 15.48 20.3022 19.6292 17.7542 22.1772C17.2765 22.6549 16.7977 22.6539 16.4779 22.3341ZM24.6315 29.3729C24.313 29.6914 23.7546 29.77 23.3548 29.3702C22.955 28.9704 22.954 28.4917 23.3521 28.0935C29.085 22.3606 29.0652 12.9451 23.3081 7.188C22.9882 6.86816 22.9073 6.30945 23.3054 5.91133C23.7035 5.5132 24.1822 5.51421 24.582 5.91402C30.9792 12.4704 31.081 22.9234 24.6315 29.3729ZM13.9826 12.0357C13.983 12.1952 13.9037 12.4344 13.7444 12.5937C13.5852 12.7529 13.346 12.8322 13.1864 12.8319L10.633 12.8265C9.67554 12.8245 8.8793 13.6207 8.88132 14.5782L8.65388 20.243C8.65589 21.2005 9.4555 22.0001 10.413 22.0021L13.2855 22.0082C13.7643 22.0092 14.1641 22.409 14.1651 22.8877L14.11 34.617C14.1103 34.7766 14.0311 35.0158 13.8718 35.175C13.7126 35.3343 13.4734 35.4136 13.3138 35.4132C13.1542 35.4129 9.48375 35.4052 8.60586 35.3235C6.37168 35.3188 3.89359 33.1592 2.69082 30.364C0.0447299 24.2144 0.499944 13.0445 2.39848 5.22894C2.55672 4.59094 3.27233 3.39557 4.22781 2.44008C5.10368 1.56422 6.53758 0.450156 8.61184 0.294943C10.5265 0.139393 13.0002 0.224397 13.1598 0.224733C13.6386 0.225742 14.0384 0.625542 13.9594 1.02433L13.9826 12.0357ZM8.77513 2.05071C7.25924 2.1273 6.22379 3.00283 5.50718 3.71945C4.63131 4.59531 4.15458 5.55181 4.15492 5.71139C1.93923 14.4838 1.96041 24.5376 4.28532 29.729C5.32749 32.0452 7.40545 33.6454 8.6025 33.7277C9.24084 33.729 11.1558 33.7331 12.4325 33.7358L12.4117 23.8415L10.3371 23.8372C8.42207 23.8331 6.90282 22.3139 6.89879 20.3989L7.12622 14.7341C7.12253 12.9787 8.71501 11.3862 10.5501 11.3103L12.3055 11.314L12.286 2.0581C11.3281 1.8965 9.89188 1.89347 8.77513 2.05071Z"
-                    fill="#0C335E"
-                  />
-                </svg>
-
-                <li class="flex items-center">+01 234 567 88</li>
-              </span>
-              <span className="flex">
-                <svg
-                  width="30"
-                  height="36"
-                  viewBox="0 0 30 36"
-                  fill="none"
-                  xmlns="http://www.w3.org/2000/svg"
-                >
-                  <path
-                    d="M19.6454 10.8507C19.3255 10.5309 19.2445 9.97215 19.6427 9.57403C20.0408 9.17591 20.5195 9.17692 20.9193 9.57672C25.3172 13.9745 25.3323 21.1558 20.953 25.5352C20.6345 25.8537 20.0761 25.9323 19.6763 25.5325C19.2765 25.1327 19.2755 24.6539 19.6736 24.2558C23.4163 20.673 23.4035 14.6088 19.6454 10.8507ZM16.4779 22.3341C16.158 22.0142 16.0771 21.4555 16.4752 21.0574C18.3862 19.1464 18.3798 16.1143 16.4607 14.1953C16.1409 13.8754 16.0599 13.3167 16.458 12.9186C16.8562 12.5205 17.3349 12.5215 17.7347 12.9213C20.2934 15.48 20.3022 19.6292 17.7542 22.1772C17.2765 22.6549 16.7977 22.6539 16.4779 22.3341ZM24.6315 29.3729C24.313 29.6914 23.7546 29.77 23.3548 29.3702C22.955 28.9704 22.954 28.4917 23.3521 28.0935C29.085 22.3606 29.0652 12.9451 23.3081 7.188C22.9882 6.86816 22.9073 6.30945 23.3054 5.91133C23.7035 5.5132 24.1822 5.51421 24.582 5.91402C30.9792 12.4704 31.081 22.9234 24.6315 29.3729ZM13.9826 12.0357C13.983 12.1952 13.9037 12.4344 13.7444 12.5937C13.5852 12.7529 13.346 12.8322 13.1864 12.8319L10.633 12.8265C9.67554 12.8245 8.8793 13.6207 8.88132 14.5782L8.65388 20.243C8.65589 21.2005 9.4555 22.0001 10.413 22.0021L13.2855 22.0082C13.7643 22.0092 14.1641 22.409 14.1651 22.8877L14.11 34.617C14.1103 34.7766 14.0311 35.0158 13.8718 35.175C13.7126 35.3343 13.4734 35.4136 13.3138 35.4132C13.1542 35.4129 9.48375 35.4052 8.60586 35.3235C6.37168 35.3188 3.89359 33.1592 2.69082 30.364C0.0447299 24.2144 0.499944 13.0445 2.39848 5.22894C2.55672 4.59094 3.27233 3.39557 4.22781 2.44008C5.10368 1.56422 6.53758 0.450156 8.61184 0.294943C10.5265 0.139393 13.0002 0.224397 13.1598 0.224733C13.6386 0.225742 14.0384 0.625542 13.9594 1.02433L13.9826 12.0357ZM8.77513 2.05071C7.25924 2.1273 6.22379 3.00283 5.50718 3.71945C4.63131 4.59531 4.15458 5.55181 4.15492 5.71139C1.93923 14.4838 1.96041 24.5376 4.28532 29.729C5.32749 32.0452 7.40545 33.6454 8.6025 33.7277C9.24084 33.729 11.1558 33.7331 12.4325 33.7358L12.4117 23.8415L10.3371 23.8372C8.42207 23.8331 6.90282 22.3139 6.89879 20.3989L7.12622 14.7341C7.12253 12.9787 8.71501 11.3862 10.5501 11.3103L12.3055 11.314L12.286 2.0581C11.3281 1.8965 9.89188 1.89347 8.77513 2.05071Z"
-                    fill="#0C335E"
-                  />
-                </svg>
-
-                <li class="flex items-center">+01 234 567 89</li>
-              </span>
-            </ul>
-          </div>
-        </div>
-        <div class="bg-blue-700 text-center text-white py-4 mt-4">
-          <p>&copy; 2020 Copyright: MDApplora.com</p>
-        </div>
-      </footer>
-    </section>
-  );
-}
-
-export default Footer;
