--- src/components/Sections/Navigation/navigation.scss
@@ -1,6 +1,0 @@
-@media screen and (max-width: 768px) {
-    #menu-items {
-        display: none;
-    }
-
-}
