--- src/components/Sections/section-styles.scss
@@ -1,92 +1,0 @@
-@import "../../Styles/mixins.scss";
-
-h1,
-h2 {
-    font-size: 2.5rem !important;
-    line-height: 1.25;
-    font-weight: 700 !important;
-}
-
-@media screen and (max-width: 1020px) {
-
-    h1,
-    h2,
-    h3 {
-        font-size: 1.8rem !important;
-    }
-
-}
-
-.ul-list-img {
-    li {
-        padding: 3px 0px 3px 0px;
-        margin-bottom: 2%;
-        border: 2px solid blue;
-        align-content: center;
-        align-self: center;
-
-    }
-}
-
-@media screen and (max-width: 1020px) {
-    .ul-list-img {
-        li {
-            img {
-                width: 15px;
-                height: 15px;
-                margin-top: 7px;
-            }
-        }
-    }
-}
-
-
-.center-text {
-    text-align: center;
-    display: flex;
-    justify-content: center;
-
-    &-div {
-        width: 40%;
-        border: 2px solid red;
-    }
-}
-
-.flex-div {
-    border: 2px solid red;
-    display: flex;
-
-    &-one {
-        width: 60%;
-        border: 2px solid black;
-    }
-
-    &-two {
-        width: 40%;
-        border: 2px solid black;
-    }
-}
-
-@media screen and (max-width: 1020px) {
-    .flex-div {
-        display: block;
-
-
-        &-one {
-            width: auto !important;
-        }
-
-        &-two {
-            width: auto !important;
-        }
-    }
-
-
-}
-
-
-
-.time-tracking {
-    background-image: url(../../assets/images/TimeTracking.svg);
-
-}
